Add a transaction datatype
The `Transaction` will be the core atomic data type for representing
acquiring or disposing of a particular crypto. It has ticker, price (in
USD), date, amount, and info on what type of transaction it is. There
are helper functions with tax relevant info based on the type (is it
generating capital gains, is it income or loss).

Note that crypto-to-crypto trades are very common, and those are
represented by two transactions: one transaction that sells the source
crypto for USD, and one transaction that buys the destination crypto for
USD. From a tax compliance standpoint, this is the right way to do it,
and it has the addon benefits of allowing all the core tx processing
code to just handle a single datatype. There is no need for an explicit
enum distinguishing buys from sells, because it is obvious from looking
at the amount (negative amount implies sell, positive amount implies
buy).

In a followon PR, I intend to make the CostBasisCalculator and
CapitalGainsCalculator use transactions rather than taking a subset of
the transaction fields as function arguments. I expect this to be more
readable (don't need to remember which is price and which is amount),
and make the codebase more consistent.

Test plan:
The implementation of this PR is quite trivial. I added tests that
verify that every transaction type is handled by the transaction
categorizers, and that the categorizers would error if provided an
unexpected tx type. I didn't bother testing the outputs of those
functions due to their trivial implementations.