Add a Transaction datatype.
It is a standard way of representing acquiring or disposing of some
cryptoasset. It uses bigjs arbitrary-precision numeric representations
for the cryptocurrency amount and price.

There are a number of transaction types, such as "TRADE", "FEE", and
"DIV". Each transaction type is associated with a TransactionTypeInfo,
with information on the tax implications of that kind of transaction.

Test plan:
The tests thus far all focus on typechecking. Observe that `yarn flow`
and `yarn test` both pass at present.

If you uncomment the invalid transaction in transaction.test.js, then
`yarn flow` produces errors. Also, if you make the transactionTypes not
all map to a TransactionTypeInfo (e.g. change isCapitalGains to "foo"
for some tx type), then `yarn flow` fails in the static assertion.