Add LIFOCostBasisCalculator and tests (#1)
The LIFOCostBasisCalculator keeps track of a sequence of acquisitions
and disposals for a given cryptocurrency. For each disposal, it reports
a sequence of `BasisFrames`; these BasisFrames have the cost basis
information for that disposal (when the corresponding crypto was
acquired, and at what cost).

Test plan: Check out the new tests that have been added. They pass, as
does `yarn flow`.